[fix] use singleton-based logger
Refs: #11
